# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Import XML API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Server API Rest"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "get token",
                "operationId": "App\\Http\\Controllers\\LoginController::index",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Token",
                                    "title": "Token",
                                    "description": "User Token",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/people": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get all people",
                "operationId": "App\\Http\\Controllers\\PeopleController::all",
                "responses": {
                    "200": {
                        "description": "People",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "People",
                                    "title": "People",
                                    "description": "People",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Person"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/people/{id}": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get person by id",
                "operationId": "App\\Http\\Controllers\\PeopleController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "People",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Person",
                                    "title": "Person",
                                    "description": "Person",
                                    "properties": {
                                        "person": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/shiporders": {
            "get": {
                "tags": [
                    "ShipOrder"
                ],
                "summary": "Get all shiporders",
                "operationId": "App\\Http\\Controllers\\ShipordersController::all",
                "responses": {
                    "200": {
                        "description": "Shiporders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Shiporders",
                                    "title": "Shiporders",
                                    "description": "Shiporder",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Shiporder"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/shiporders/{id}": {
            "get": {
                "tags": [
                    "ShipOrder"
                ],
                "summary": "Get shiporder by id",
                "operationId": "App\\Http\\Controllers\\ShipordersController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ShipOrder ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shiporders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Shiporder",
                                    "title": "Shiporder",
                                    "description": "Shiporder",
                                    "properties": {
                                        "shiporder": {
                                            "$ref": "#/components/schemas/Shiporder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "App\\Http\\Controllers\\UserController::all",
                "responses": {
                    "200": {
                        "description": "User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "User",
                                    "title": "User",
                                    "description": "User",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "create user",
                "operationId": "App\\Http\\Controllers\\UserController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Token",
                                    "title": "Token",
                                    "description": "User Token",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "operationId": "App\\Http\\Controllers\\UserController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "User",
                                    "title": "User",
                                    "description": "User",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Person": {
                "title": "Person",
                "description": "Person Imported",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "shiporders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Shiporder"
                        }
                    }
                },
                "type": "object"
            },
            "Phone": {
                "title": "Phone",
                "description": "Phone of person",
                "properties": {
                    "number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Shiporder": {
                "title": "ShipOrder",
                "description": "Ship Order from Person",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "shipto_address": {
                        "type": "string"
                    },
                    "shipto_name": {
                        "type": "string"
                    },
                    "shipto_city": {
                        "type": "string"
                    },
                    "shipto_country": {
                        "type": "string"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Items"
                        }
                    }
                },
                "type": "object"
            },
            "Items": {
                "title": "Items",
                "description": "Items of shiporder",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "float"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User login",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}